LayoutTests/inspector-obsolete/http-tests/network/network-clear-cookies-expected.txt

CONSOLE MESSAGE: line 13: First XHR loaded.
CONSOLE MESSAGE: line 28: Done.
Tests clearing browser cookies from inspector.

http://127.0.0.1:8000/inspector/network/resources/set-cookie.php
http://127.0.0.1:8000/inspector/network/resources/set-cookie.php
